Get StreamCat Metric Names

Function to filter StreamCat metrics metrics by category, area of interest, dataset or year. Use `sc_get_params(categories)` or `sc_get_params(datasets)` to see all the valid category or dataset options

Usage

sc_get_metric_names(category = NULL, aoi = NULL, year = NULL, dataset = NULL)

Arguments

category

Filter StreamCat metrics based on the metric category

aoi

Filter StreamCat metrics based on the area of interest

year

Filter StreamCat metrics based on a particular year or years

dataset

Filter StreamCat metrics based on the dataset name

Value

A dataframe of merics and description that match filter criteria

Author

Marc Weber

Examples

if (FALSE) { # \dontrun{
metrics <- sc_get_metric_names(category='Wildfire')
metrics <- sc_get_metric_names(category = c('Deposition','Climate'),
aoi=c('Cat','Ws'))
metrics <- sc_get_metric_names(aoi='Other',
dataset=c('Canal Density','Predicted Channel Widths Depths'))

} # }