Get StreamCat data

Function to return StreamCat catchment and watershed metrics using the StreamCat API. The function allows a user to get specific metric data aggregated by area of interest, returned by comid(s), hydroregion(s), state(s), or county(ies).

Usage

sc_get_data(
  metric = NULL,
  aoi = NULL,
  comid = NULL,
  state = NULL,
  county = NULL,
  region = NULL,
  showAreaSqKm = NULL,
  showPctFull = NULL,
  conus = NULL,
  countOnly = NULL
)

Arguments

metric

Name(s) of metrics to query Syntax: name=<name1>,<name2>

aoi

Specify the area of interest described by a metric. By default, all available areas of interest for a given metric are returned. Syntax: areaOfInterest=<value1>,<value2> Values: catchment|watershed|riparian_catchment|riparian_watershed|other

comid

Return metric information for specific COMIDs. Needs to be a character string and function will convert to this format if needed. Syntax: comid=<comid1>,<comid2>

state

Return metric information for COMIDs within a specific state. Use a state's abbreviation to query for a given state. Syntax: state=<state1>,<state2>

county

Return metric information for COMIDs within a specific county. Users must use the FIPS code, not county name, as a way to disambiguate counties. Syntax: county=<county1>,<county1>

region

Return metric information for COMIDs within a specified hydroregion. Syntax: region=<regionid1>,<regionid2>

showAreaSqKm

Return the area in square kilometers of a given area of interest. The default value is false. Values: true|false

showPctFull

Return the pctfull for each dataset. The default value is false. Values: true|false

conus

Return all COMIDs in the conterminous United States. The default value is false. Values: true|false

countOnly

Return a CSV containing only the row count (ROWCOUNT) and the column count (COLUMNCOUNT) that the server expects to return in a request. The default value is false. Values: true|false

Value

A tibble of desired StreamCat metrics

Author

Marc Weber

Examples

# \donttest{
df <- sc_get_data(comid='179', aoi='catchment', metric='fert')
#> Rows: 1 Columns: 4
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl (4): COMID, WSAREASQKM, CATAREASQKM, FERTCAT
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

df <- sc_get_data(metric='PctGrs2006', aoi='watershed', region='01')
#> Rows: 65968 Columns: 4
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (1): REGIONID
#> dbl (3): COMID, WSAREASQKM, PCTGRS2006WS
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

df <- sc_get_data(metric='PctUrbMd2006', aoi='riparian_catchment',
comid='1337420')
#> Rows: 1 Columns: 4
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl (4): COMID, WSAREASQKMRP100, CATAREASQKMRP100, PCTURBMD2006CATRP100
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

df <- sc_get_data(metric='PctUrbMd2006,DamDens',
aoi='catchment,watershed', comid='179,1337,1337420')
#> Rows: 3 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl (7): COMID, WSAREASQKM, CATAREASQKM, PCTURBMD2006CAT, PCTURBMD2006WS, DA...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

df <- sc_get_data(metric='PctUrbMd2006,DamDens',
aoi='catchment,watershed', comid='179,1337,1337420',
showAreaSqKm=FALSE, showPctFull=TRUE)
#> Rows: 3 Columns: 10
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl (10): COMID, WSAREASQKM, PCTURBMD2006CAT, PCTURBMD2006WS, nlcd2006.CatPc...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

df <- sc_get_data(metric='PctUrbMd2006,DamDens',
aoi='catchment,watershed', comid='179,1337,1337420', countOnly=TRUE)
#> Rows: 1 Columns: 2
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl (2): ROWCOUNT, COLUMNCOUNT
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.

 # }